requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
